refactor: 유틸 함수 분리 - 헬퍼 함수
