+++ src/basic/utils/index.js
@@ -1,3 +1,4 @@
 // 🔥 Barrel exports - 모든 유틸리티를 한 곳에서 import 가능
 export * from './calculations.js';
-export * from './ui.js'; 
+export * from './ui.js';
+export * from './helpers.js';
